feature: prefer LAGO_FRONT_URL and enable urls with non-standard ports
Before this change you had to manually edit this file in order to run
the api and front-end on a non-standard port in development. Now
LAGO_FRONT_URL will be prefered if it exists in the environment and it
will allow non-standard ports instead of just using the configured
hostname.