feat(rbac): Add role to user invitation GQL endpoints and services (#1969)
## Context

Granular permissions are being added to Lago.

## Description

When inviting a user to an app, you must assign them a role. This PR
adds role to user invitation endpoints (create and accept) and related
services.

`Invite.role` attribute is added and on par with `Membership.role`.

I removed the email param when possible, instead of using the email
passed to graphql, we use the email from the invite we retrive via the
token.