feat(fees): Add billing_at arg to CreatePayInAdvance flow (#2087)
## Context

When a charge is `invoiceable: true, pay_in_advance: true` an invoice is
created every time an event is received. If the Billable Metrics
attached is `recurring: true`, each month this charge will be added to
the subscription invoice.

If the charge is `invoiceable: false, pay_in_advance: true` a fee is
created every time an even is received (but no invoice). If the Billable
Metrics attached is `recurring: true`, nothing happens each months.

We want to change this behavior and create a new Fee each month.

## Description

Each time a new fee is created for the new period, we'll continue to
attach it to the same event that triggered it in the first place.

In this PR, we extract the "timestamp" into an arg so on following
months, we can use the correct billing date instead of relying on the
time of the event.

Without this change, the boundaries of any fee would always be the
boundaries computed at the time of the event.