fix(service): Ignore current_user if it's not an instance of User (#2134)
## Description

All services inherit from `BaseService`. Many services override the
`initialize` constructor method and call `super`.

The BaseService constructor only accepts one argument: the current user.
When we call `super` without parenthesis, the argument are forwarded to
the parent. Most of the time, we call the parent constructor without the
parenthesis.

It means that in many many case, the result returned by the service has
a `result.user` which old something else, not a user.
We heavily use the named parameters so most of the time, it's a hash
with the argument of the descendant service.

One way to fix this is to go and modify all calls to `super` and replace
them with `super(nil)`. I think a better way is to ignore the
current_user if it's not an instance of `User`.

### Example

```ruby
module AddOns
  class DestroyService < BaseService
    def initialize(add_on:)
      @add_on = add_on
      super
    end

    # ...

  end
end
```

![CleanShot 2024-06-05 at 14 43
16@2x](https://github.com/getlago/lago-api/assets/1525636/ade737ff-8978-406c-9919-60425c3b35b0)