feat(pdf): Ensure Invoice PDF is created before sending emails and webhooks (#2121)
### Context
Separate PDF generation into a different Worker and Queue (:pdf). 

### Description
Move PDF generation to a distinct worker and queue.  
Change the process to generate the PDF and send the email 


> [!CAUTION]
> This PR introduces a new sidekiq queue named `pdf`.
> Other than the sidekiq config here, what else should be updated?


### Changes Introduced

- **New Job Implementation**: Introduced a new job
`GeneratePdfAndNotifyJob` to handle the PDF generation and email
notification.
- **Mailer Invocation Update**: Modified the invoicing process to
utilize the new job, ensuring the email is sent only after the PDF is
successfully generated.
- **Test Configuration**: Added a `PdfHelper` module and updated RSpec
configuration to stub PDF generation in tests tagged with `:scenarios`,
reducing repetitive stubbing code in individual tests.

## Webhook Events

> [!IMPORTANT]  
> Stop listening to `invoice.created` to request the PDF creation via
the API. Instead, listen to `invoice.generated` event, which always has
the PDF file url.

### `invoice.created` webhook

This webhook is sent after the invoice object is created, in the webhook
payload you get an `invoice` object but the PDF link is not set yet. The
PDF was not generated yet.

Until now, users would typically call back our API to request the PDF
generation. **This is now deprecated!**

Instead, wait for the `invoice.generated` webhook.

### `invoice.generated` webhook

This webhook is sent when the invoice PDF is generated. Until now, we
generated the PDF only if you requested it via the API, via the UI or if
the invoice was sent over email. If you don't send invoice over email,
you would rarely receive this event.

Now, **the PDF is generated every time an Invoice model is created**.
It's now recommended to listent this event if you want the PDF file URL
to be part of the webhook payload.

| Webhook name | Desc |
|----------------:|--------|
| `invoice.created` | This webhook indicates that the invoice has been
created, but the PDF is not yet ready. It's delivered fast after the
invoice creation. |
| `invoice.generated` | This webhook indicates that the PDF has been
generated. You receive the Invoice object WITH the PDF URL. This webhook
is sent a bit after `invoice.created` because generating PDF is
typically a bit longer. |

---------

Co-authored-by: Miguel Pinto <darkymiguel@gmail.com>