feat(boundaries): Move charge_pay_in_advance_interval from Invoice to DatesService (#2193)
## Description

For Pay in Advance charges, the date boundaries are sometimes
incorrectly computed. This was patch recently to ensure the dates show
in the invoice, in the frontend and in the API are correct but the
database is still not-so-correct.

Ideally, we would want to fix the boundaries and store the dates in each
fees but it's a significant effort and I cannot do it now. This PR
extract the previous patch from `Invoice` model to `DatesService` so I
can reuse it for [Non Invoiceable
fees](https://github.com/getlago/lago-api/pull/2185).

See these two PRs for more context:
https://github.com/getlago/lago-api/pull/1188 and
https://github.com/getlago/lago-api/pull/1540