feat(fees): Only allow payment status to be changed if not attached to invoice yet (#2197)
## Description

When [non-invoiceable fees were
introduced](https://github.com/getlago/lago-api/pull/1069), we only
allowed change of `payment_status` if the fee was attached to a
non-invoiceable charge.

What we really meant is _"the fee is standalone, not attached to an
invoice"_. Now we're working on attaching the non-invoiceable fees to an
invoice at the end of the period. See
https://github.com/getlago/lago-api/pull/2171

I'm changing the condition so the fee can be updated, only if it's not
attached to any invoice yet.