feat (integration-customers): refactor integration customers handler (#2196)
## Context

Currently Lago is adding accounting and tax integrations

## Description

Logic for integration customers needed refactoring.

Logic for integration customers removal was based on the fact that
`external_customer_id` is nil or not. However, with anrok integration,
this logic is not sufficient since integration customer object can exist
without `external_customer_id` (because integration customer is synced
with the first invoice and not right away).

That being said, we added logic that is already applied on metadata
objects and that can be applied to cover described case. Objects that
are not sent in arguments are removed.