feat(lib): Move VersionService to LagoUtils and initialize at boot time (#2204)
## Description

Currently, the Lago version resolver is a regular `Service`:
* Every new call to the service will resolve the version again (read the
file or its metadata)
* There is a `user` set to nil in the OpenStruct returned 

It's fine because we don't use this service so often. I'm going to add
the Lago version to the metadata stored by papertrail so it's going to
be called a lot more. It will be done separately.

With this PR, the service is moved to the `LagoUtils` module under the
`lib` directory. It's computed once when Rails is booted via an
initializer. In the app code, we'll now find the OpenStruct in the
`LAGO_VERSION` constant.