feat (integration-customers): add external account id to the anrok GQL response (#2209)
## Context

Currently multiple tax and accounting integration are being added to the
Lago app

## Description

This PR adds `externalAccountId` to the GQL response of the integration
object.

This field is is needed for the UI to build the URL link for anrok
customer. We can extract this data from api key but we don't want to
expose whole api key object so new field in the response is added.