feat(fees): Invoice non-invoiceable fees at the end of the period (#2171)
## Overview

This pull request introduces a new feature to handle fees paid in
advance that are not invoiceable. The goal is to provide clear
visibility for end users on what they have paid during a specific
billing period and to ensure all such fees are grouped into invoices at
the end of each billing period.

## Changes Introduced

> 
> The code in this PR is currently commented out because we are planning
to merge another PR that will allow this code to be toggled based on a
setting. Once the subsequent PR is merged, this code will be enabled
accordingly.

Feel free to adjust the wording to better fit your style or to provide
additional context if necessary.

- **New Job Implementation**: Introduced `BillNonInvoiceableFeesJob` to
handle the periodic billing of non-invoiceable fees, ensuring they are
invoiced at the end of each billing period.
- **New Invoice Type**: Added a new invoice type `advance_charges` to
categorize and manage the fees paid in advance that were previously not
included in invoices.
- **Service for Advance Charges**: Created
`Invoices::AdvanceChargesService` to generate invoices for
non-invoiceable fees, ensuring all relevant fees within the billing
period are grouped and included.
- **Integration in Billing Workflows**: Integrated the new job and
service into existing billing workflows within
`Subscriptions::BillingService`, `Subscriptions::CreateService`, and
`Subscriptions::TerminateService` to ensure seamless operation and
timely invoicing.
   
## New Invoice template
   
![CleanShot 2024-06-27 at 11 13
56@2x](https://github.com/getlago/lago-api/assets/1525636/dd14317d-ced2-4e46-8c2e-47a853bf75e9)

---------

Co-authored-by: Miguel Pinto <darkymiguel@gmail.com>