feat(event): Add view to allow caching of charge/filters for billable_metrics (#2231)
## Context

This PR is part of the real-time aggregation epic

## Description

This PR adds a new `billable_metrics_grouped_charges` database view. It
selects the charges for each organization/billable metric code/plan
tupple. It will be used to route events received on the ingest service
to the right bucket in order to perform a pre-computation and allow
future "real time aggregation"