feat(charges): Add `regroup_paid_fees` to Charge (#2232)
## Context

When your charge is pay in advance but not invoiceable, the fees are
created but not attached to any invoice.
This adds an options so at the end of the period, you get a new invoice
showing all PAID fees. unpaid fees are ignored.

See also https://github.com/getlago/lago-api/pull/2171

## Description

We're dropping the idea of `invoice_strategy` because it turned out to
be a bad idea. Instead we're adding `regroup_paid_fees` attribute that
takes only one value today `invoice`.