fix(export-invoices): allow no pagination for Query objects (#2307)
## Roadmap Task

👉
https://getlago.canny.io/feature-requests/p/ability-to-export-data-from-the-user-interface

## Context

When exporting data (eg: invoices) we use query objects to fetch the
records to export. However, the current implementation of query objects
paginates to first page and limit results to 25 or 100 (depends on
implementation) when no pagination information is provided.

We need the query object to allow no pagination and fetch all resources
matching the search / filter criteria when pagination args are not provided.

## Description

This change allows client code to call Query object with `pagination: nil`
arg to avoid pagination defaults (page: 1, limit: 25).

Also, it makes `BaseQuery::Pagination` internal knowledge of the query
objects. As query objects expect pagination as a hash, for example: `{page: 1, limit:
50}`. The same will be applied to `BaseQuery::Filters` in a coming PR.

Updates the Query objects implementing the latest approach and updates
`InvoicesQuery` object to allow no pagination and fix the CSV data
export issues.