chore(services): Rename FinalizeService to RefreshDraftAndFinalizeService (#2302)
## Description

As we're considering adding more Invoice state, we should specify what
the current `FinalizeService` does. It not only finalize it, it takes a
draft invoice, refresh it and finalize it.

We're also changing a little bit how invoice is set in result and
reloaded, hoping to make it a little bit clearer.