Fix InvoicesQuery filters to ensure it works from all kind of usage...
from the graphQL invoices resolver
from the REST API invoices controller
and from the DataExports services given the filters are saved in a
database fields and keys are converted into strings.

Also, this change fixes the query performed for the CSV export as before
was missing some filters that currently are outside of the filters hash.