Fix InvoicesQuery payment dispute lost filter
Whenever `payment_dispute_lost` filter had a value other than `nil` it
was returing only inovoices with `payment_dispute_lost`,

hence, when the filter had a `false` value, it still returns only
invoices with a payment dispute lost.

This change fixes the filter by providing the proper scoping base on the
value of the filter.