feat(wallet): Add `invoice_require_successful_payment` attributes on wallet-related models (#2326)
## Description

We're introducing a way to topup your wallet without an invoice. The
invoice is only generated after the payment is made.

Internally, when the wallet transaction is initialized, we'll create [an
`open` invoice](https://github.com/getlago/lago-api/pull/2330), which is
invisible and doesn't have a number yet. This invoice can be paid. Once
it's paid it's going to be finalized. (see #2336)

This PR introduce a parameter to control this behavior.

* When a `WalletTransaction` is created, the
`invoice_require_successful_payment` attribute can be set explicitly.
 * If not set, it follows the `Wallet` configuration
* When a wallet `RecurringTransactionRule` is created, the
`invoice_require_successful_payment` attribute can be set explicitly.
  * If not set, it follows the `Wallet` configuration
* When a `WalletTransaction` is created from a
`RecurringTransactionRule`, the invoice_require_successful_payment` is
always read from the rule.
* All three models default to `false` in the DB