Fix CSV file output records order by issuing date, newest on top
While find in batches is the performance optimal solution it does not
allow custom ordering, just primary key ordering. Then, as CSV export
runs as a background job we will test the impact with the easiest
solution. We will paginate the CSV exporter if retrieving records impacts
worker perfromance or memory consumption.