fix(netsuite): Do not send taxes columns if tax not mapped (#2369)
## Context

In our NetSuite integration, if the tax_item mapping is not set, do not
block the invoice from being synced to NetSuite. In that case, simply
remove the taxitem and the taxamountoverride from the POST invoice
request.

## Description

Remove`taxitem` and `taxamountoverride` columns from invoice payload if
tax item is not mapped.