feat(ProgressiveBilling) -  add `LifetimeUsages::CalculateService` (#2401)
## Context

AI companies want their users to pay before the end of a period if usage
skyrockets. The problem being that self-serve companies can overuse
their API without paying, triggering lots of costs on their side.

## Description

This PR add the `LifetimeUsages::CalculateService`. This service is
responsible for correctly calculating the `invoiced_usage` and
`current_usage` amounts.

Secondly a small refactor to the `CustomerUsageService` to differentiate
between the 2 use-cases and streamline usage in the newly added
`CalculateService`