Fix(anrok): no fallback_item mapping is found (#2428)
## Context

Now if user doesn't have a fallback item set up and tries to fetch taxes
for item without product_mapping provided, we'd fail with `undefined
method 'external_id' for nil (NoMethodError)` and will return server
error. Instead we should process it as any other tax_integration error,
because it's an incomplete integration set up

## Description

in case when no mapped_item is found for the fee, we provide an empty
open_struct, so calling any method on it returns nil. It allows us to
send request to anrok with empty external_product_id, and then anrok can
return us validation error with message `"Request body: \"lineItems\":
0: \"productExternalId\": String must contain at least 1 character(s)."`
which indicates, that product_external_id to fetch taxes is missing