feat(ProgressiveBilling) - Add historical_usage_amount_cents to LifetimeUsage (#2457)
## Context

There's always a risk of customers not paying invoices generated at the
end of a billing period. It would be beneficial to bill customers at
given thresholds (units vs. amount) rather than waiting until the end of
the period. This approach would allow for removing customer access if
invoices are not paid and prevent having a highest amount of unpaid
invoices.

## Description

Add a `historical_usage_amount_cents` column to `LifetimeUsage`, this
column can be filled in with usage coming from an external system.

We take the historical usage into account when checking if thresholds
have been passed.