feat(ProgressiveBilling) - Add support for LAGO_LIFETIME_USAGE_REFRESH_INTERVAL_SECONDS  (#2460)
## Context

There's always a risk of customers not paying invoices generated at the
end of a billing period. It would be beneficial to bill customers at
given thresholds (units vs. amount) rather than waiting until the end of
the period. This approach would allow for removing customer access if
invoices are not paid and prevent having a highest amount of unpaid
invoices.

## Description

Allows configuring the refresh interval via an ENV variable.