fix(thresholds): Fix updating plan with deleted thresholds (#2479)
## Context

We were getting validation erros when updating a plan via the UI.
The problem was a specific combination of: removing and re-adding
thresholds with the same amount because of a constraint. First we were
creating a new threshold while the old one (with the same amount was
still in the DB). Also there might have been the same threshold wtih
that amount existing but soft deleted.

## Description

Refactor the validations, and saving logic of thresholds.