Feat(anrok<>credit_notes) refactor CreditNotes::AppliedTax to search taxes by tax_code (#2536)
**Context**
When we apply taxes on credit_notes, we're applying taxes that are
stored for invoice - Invoice::AppliedTax. (and then we select taxes
applied on fees, related to the credit_note_items)

Previous taxes implementation suggested that we should have a tax, saved
in the db, so when iterating through taxes we were using tax.id, but
currently taxes received from tax provider don't have a corresponding
tax in the db, but we have tax_code, identifying the tax. the same is
with stored in the db taxes: within the organization the tax_code will
not be repeated, so if we start looking for taxes by tax_code instead of
tax.id, we won't break existing logic and will adapt it to the new
applied taxes that do not have taxes, stored in the db.

**What was done:**
Updated algorythm of searching taxes to apply on credit note, to search
them by tax_code instead of tax_id