misc(cleanup) Fix billing service call & introduce rubocop ensuring code consistency (#2543)
## Description

We've built several subclasses of BaseService that don't adhere to the
common pattern of having an `initialize` method with arguments and a
`call` method without arguments. In an effort to standardize this
pattern, a rubocop is added that validates that the call method
definition on subclasses of BaseService does not have any arguments.