misc(event): Remove ordering of event queries when not required (#2618)
## Context

When running aggregation for billable metrics we are (almost) always
ordering events by the `timestamp` property.
In most cases this order is not necessary since the end-result will not
be dependent of the event order (sum, count, max...) and the order
clause might slow down the queries.

## Description

This PR updates all events-related aggregation queries to get rid of the
order and keep it only when required. This should allow for a small gain
in aggregation performances.