feat(event): Add event list API (#2601)
## Roadmap Task

👉 https://getlago.canny.io/feature-requests/p/retrieve-and-filter-events

## Context

Currently, REST API does not provide the endpoint to list all events (as
in the Developer -> Events page). This pull request aim to expose the
event list to the REST API to let the user retrieve and filter events.

## Description

This pull request add supports for:  
- Adding `GET /api/v1/events` endpoint
- Filtering events by `code`, `external_subscription_id`,
`timestamp_from`, and `timestamp_to`

I'm also planning to checkout the front-end side to add these kinds of
filter into the UI later as well.