misc: Refact some services to the call pattern (#2641)
## Context

Some services inherinting from the `BaseService` are not using the
`SERVCE_CLASS.call` pattern. They are mostly old services that were
created before the introduction of the pattern in the project.

## Description

To increase consistency in the code base, this PR turns some old
services to the pattern.
The impacted services are:
- `AddOns::CreateService`
- `BillableMetrics::CreateService`
- `Coupons::CreateService`
- `AddOns::CreateService`

More refactor will follow as other services have still to be converted.