Feat: Add price breakdown for pay in advance percentage charge (#2605)
## Context

We needed to enhance representation of price for percentage pay in
advance charge. Initial request was to simply show only percentage if
only percentage is set, but then the request grew up to show the whole
amount breakdown as we have in subscription invoices

## Description

- added amount_details calculation for fee as we have for fee
aggregation (subscription invoices)
- amount_detail is calculated only for percentage invoices
- added fix for free_events calculation: before if we had a plan that
allows 3 free events, we received two and generate the invoice, in the
invoice it will say that there are 0 free events. Now it will say 2
free_events