misc: Remove the need for BaseService current user argument (#2660)
## Context

The `BaseService#initialize` method is expecting a first argument of
type `User` that is then assigned to the created `result`. This argument
is not used in any places (last references were removed in previous PRs)

## Description

This PR simply removes this reference, and refactors the
`::PaymentProviders::DestroyService` that was the last remaining service
relying on it.