fix(pdf): Fix credit note generating pdf and webhook (#2659)
## Context

For credit notes, Lago does not automatically generate PDF files. The
webhook `credit_note.generated` is only sent when an API request is
made.

Expected behaviour: the credit note logic should be the same as the
invoice logic.

## Description

This PR changes logic for generating credit notes and their webhook so
that it's the same as invoices.