misc: Update cache format version (#2661)
## Context

Rails version was bumped few month ago to `7.1.X`, but to make migration
possible, the cache format version was kept to `7.0`. This was a
requirement of the standard rails upgrade process. It will also be
required to upgrade to rails `7.2` and `7.3`

## Description

Since rails 7.1 is in production for few month now it safe to change the
cache version.