feat(hubspot): Add new attributes to hubspot integration model (#2662)
## Context

The goal of this change is the deployment of custom properties and
objects in HubSpot when a new integration is created via Lago.
Specifically, we need to create the necessary fields for both native
HubSpot objects (`Companies` and `Contacts`) and create custom objects
(`Subscriptions` and `Invoices`) in the user's HubSpot account.

## Description

This PR introduces a new job, `SyncCustomObjectsAndPropertiesJob`, which
orchestrates the setup of custom objects and property synchronization in
HubSpot when a new integration is created via Lago. The job does not
handle sending the actual records, and this will be addressed in a
future PR.

version control has been implemented to track the deployment state of
these objects and properties. This allows us to know what has been
deployed and to trigger a sync when sending new records if the schema
has been updated.

---------

Co-authored-by: Ivan Novosad <ivan@getlago.com>