Feat: issue credit note on credit invoice - db change (#2676)
## Context

Now we don't allow creation of credit notes on prepaid credits. We want
to implement this possibility.
When credits will be issued, we'll be creating a voiding outbound
wallet_transaction. Currently we're tracking the outbound wallet
transactions. To support this behaviour we'll need to add credit_note_id
to the wallet transaction

## Description

Added credit_note_id to wallet_transaction