fix(clickhouse): Fix timestamp precision (#2689)
This pull request focuses on optimizing timestamp precision in the
`unique_count_query.rb` file by adjusting the precision level in various
SQL queries. The changes include modifying the `toDateTime64` function
to use a precision of 3 instead of 5 across different methods.

Timestamp precision adjustments:

*
[`app/services/events/stores/clickhouse/unique_count_query.rb`](diffhunk://#diff-15e04aaa1cc9a958e117e73b27e752ce967e85a43ce0dcb7f4ded58b44b9ff78L209-R209):
Changed the precision in the `events_cte_sql` method to use `timestamp`
directly instead of `toDateTime64(timestamp, 5, 'UTC')`.
*
[`app/services/events/stores/clickhouse/unique_count_query.rb`](diffhunk://#diff-15e04aaa1cc9a958e117e73b27e752ce967e85a43ce0dcb7f4ded58b44b9ff78L229-R229):
Updated the `grouped_events_cte_sql` method to use `timestamp` directly
instead of `toDateTime64(timestamp, 5, 'UTC')`.
*
[`app/services/events/stores/clickhouse/unique_count_query.rb`](diffhunk://#diff-15e04aaa1cc9a958e117e73b27e752ce967e85a43ce0dcb7f4ded58b44b9ff78L290-R294):
Modified the `period_ratio_sql` method to use `toDateTime64` with a
precision of 3 instead of 5 for the `from_datetime` and `to_datetime`
parameters.
*
[`app/services/events/stores/clickhouse/unique_count_query.rb`](diffhunk://#diff-15e04aaa1cc9a958e117e73b27e752ce967e85a43ce0dcb7f4ded58b44b9ff78L320-R324):
Adjusted the `grouped_period_ratio_sql` method to use `toDateTime64`
with a precision of 3 instead of 5 for the `from_datetime` and
`to_datetime` parameters.