perf(Filters) - increase performance of updating charge filters (#2696)
## Description 

Based on profiling information we found out that the `to_h` method on
ChargeFilter is called often when updating a charge with a large ( >
100) amount of filters.

A second observation is that the `touch` method means we're updating all
charge filters every time, which has poor performance when you have a
lot of those for a given charge.

## Changes

- cache `to_h` 
- updated_at is used for ordering the filters, but if you have > 100
we've chosen to optimize and not touch all records. This means stuff
will change in the UI, but at that amount nobody is really managing them
based on order.