fix(clickhouse): Uniq count should count uniq values... (#2721)
## Context

This PR fixes the unique count aggregation for clickhouse event store

## Description

The unique count implementation uses window function to avoid double
counting multiple addition or removals of the same unique property. The
clickhouse implementation was a rewriting of the one implemented in
clickhouse but it was not working as expected.